requirements.txt
requests
python-dotenv
#Commande pour les outils "pip install -r requirements.txt"